Split native Zenmap results filtering helpers
